Added the solution of Task 1
So, here is the tip.
You should run the code before submitting. I guess we all know that. but you should also change the submission language before submitting if you are using g++ or visual studio chances are you are using c++17 or c++20 by default hacker rank is set to c++ which is c++98.
Seems none sense but I have suffered a lot because of this shit.
Regards,
Arslan Iftikhar...